voicebot-11labs/requirements.txt
# Core dependencies
openai>=1.0.0
elevenlabs>=0.2.0
python-dotenv>=1.0.0

# Audio processing
pyaudio>=0.2.11
numpy>=1.24.0
scipy>=1.10.0
sounddevice>=0.4.6
pydub>=0.25.1

# Telephony and WebSocket
twilio>=8.0.0
websockets>=11.0
aiohttp>=3.8.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Speech processing
openai-whisper>=20230314
torch>=2.0.0
torchaudio>=2.0.0

# Utilities
requests>=2.31.0
asyncio>=3.4.3
colorama>=0.4.6
rich>=13.0.0

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0 